% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesc.R
\name{create}
\alias{create}
\title{Draft different templates using this function}
\usage{
create(filename, template, create_dir = TRUE)
}
\arguments{
\item{filename}{A character string specifying the filenames you like to create}

\item{template}{The template to use. If it is thesis, the filename will be used as index.Rmd}

\item{create_dir}{A logical value specifying where to create directory or not}
}
\value{
An R Markdown output format.
}
\description{
Draft different templates using this function
}
\examples{
\dontrun{
nmbutemplates::create("thesis")
}
}
