% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_question.R
\name{exam_question}
\alias{exam_question}
\title{Exam Question Template}
\usage{
exam_question(show_answer = NULL, keep_tex = TRUE, ...)
}
\arguments{
\item{show_answer}{Flag to show or hide answer}

\item{keep_tex}{Flag to keep or remove tex file}

\item{...}{Arguments passed to \code{rmarkdown::\link{pdf_document}()}.}
}
\value{
An R Markdown output format.
}
\description{
Exam Question Template
}
\section{\code{exam_question}}{
 Format for creating Exam Questions adapted from older exam formats.
}

\examples{
\dontrun{
rmarkdown::draft("STAT50.Rmd", template = "exam_question", package = "nmbutemplates")
}
}
