% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_question.R
\name{nmbutemplates}
\alias{nmbutemplates}
\title{R Markdown output formats NMBU}
\description{
Most article formats are based on \code{rmarkdown::pdf_document()}, with a
custom Pandoc LaTeX template and different default values for other arguments
(e.g., \code{keep_tex = TRUE}).
}
\section{Details}{
 You can find more details about each output format below.
}

