% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis.R
\name{thesis}
\alias{thesis}
\title{Thesis for Masters and PhD Thesis}
\usage{
thesis(...)
}
\arguments{
\item{...}{Arguments passed to \code{rmarkdown::\link{pdf_document}()}.}
}
\value{
An R Markdown output format.
}
\description{
Thesis for Masters and PhD Thesis
}
\section{\code{thesis}}{
 Format for creating masters and phd thesis
}

\examples{
\dontrun{
rmarkdown::draft("index.Rmd", template = "thesis", package = "nmbutemplates")
}
}
